+++ src/Models/CallbackForceSaveType.php
@@ -4,7 +4,7 @@ namespace Onlyoffice\DocsIntegrationSdk\Models;
 
 /**
  *
- * (c) Copyright Ascensio System SIA 2024
+ * (c) Copyright Ascensio System SIA 2025
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,6 +26,7 @@ class CallbackForceSaveType extends BasicEnum
     const COMMAND_SERVICE = 0;
     const SAVE_BUTTON = 1;
     const TIMER = 2;
+    const SUBMIT_FORM = 3;
 
     public function __construct($type = null)
     {